Change type Integer -> Long for id fields in entities classes and in dao

+++ src/main/java/com/ss/touragency/dao/CountryDao.java
@@ -48,7 +48,7 @@ public class CountryDao implements ICrudDao<Country> {
 
             while (resultSet.next()) {
                 Country country = new Country();
-                country.setIdCountry(resultSet.getInt("idCountry"));
+                country.setIdCountry(resultSet.getLong("idCountry"));
                 country.setCountryName(resultSet.getString("countryName"));
 
                 countryList.add(country);
@@ -58,7 +58,7 @@ public class CountryDao implements ICrudDao<Country> {
     }
 
     @Override
-    public Country selectById(int id) throws SQLException {
+    public Country selectById(Long id) throws SQLException {
 
         String sql = "SELECT * FROM COUNTRY WHERE idCountry=" + "'" + id + "'";
         Statement statement = DBConnection.getDbConnection().createStatement();
@@ -70,7 +70,7 @@ public class CountryDao implements ICrudDao<Country> {
         while(resultSet.next()){
             country = new Country();
 
-            country.setIdCountry(resultSet.getInt("idCountry"));
+            country.setIdCountry(resultSet.getLong("idCountry"));
             country.setCountryName(resultSet.getString("countryName"));
         }
